requirements.txt
fastapi~=0.111.0
sqlmodel~=0.0.19
pydantic~=2.8.0
python-decouple~=3.8
passlib[bcrypt]~=1.7.4
PyJWT~=2.8.0
uvicorn~=0.30.1
SQLAlchemy~=2.0.31
aiosqlite~=0.20.0